requirements.txt
streamlit
langchain
langchain-openai
langchain-community
openai
wikipedia
python-dotenv  # optional if local dev
